@@ -5,7 +5,7 @@ module StepHelper
   #the tutorials are assigned to a controller and an action
   #eventualmente, lo step da mostrare 
   def get_next_step(user=current_user)
-    tutorial_assignee = current_user.todo_tutorial_assignees.first(:joins => :tutorial, :conditions => "tutorials.action = '#{params[:action]}' and tutorials.controller = '#{params[:controller]}'", :readonly => false)
+    tutorial_assignee = current_user.todo_tutorial_assignees.joins(:tutorial).where("tutorials.action = '#{params[:action]}' and tutorials.controller = '#{params[:controller]}'").first
     check_tutorial_status(tutorial_assignee) if tutorial_assignee
   end
 
