@@ -1,6 +1,5 @@
-class Document < ActiveRecord::Base
-  unloadable
-  
+class Folder < ActiveRecord::Base
+
   #has_many :blog_posts, :through => :blog_post_images, :dependent => :destroy
 
   #attr_accessor :random_id
