@@ -8,7 +8,7 @@ module Frm
     validates :viewable_id, :viewable_type, :presence => true
     attr_accessible :user, :current_viewed_at, :count
 
-    scope :topics,  {conditions: "viewable_type = 'Frm::Topic'"}
+    scope :topics,  -> {where "viewable_type = 'Frm::Topic'"}
     def viewed_at
       updated_at
     end
