@@ -1,6 +1,6 @@
 class NotificationCategory < ActiveRecord::Base
 #  translates :description
-  has_many :notification_types, :class_name => 'NotificationType', :order => :id
+  has_many :notification_types, -> {order('id')}, :class_name => 'NotificationType'
 
   def description
     I18n.t("db.#{self.class.class_name.tableize}.#{self.short}.description")
