@@ -17,7 +17,8 @@ class SysFeature < ActiveRecord::Base
     data-number=\"#{self.id}\"
     data-atype=\"feature\"
     data-quantity=\"1\"
-    data-env=\"#{Rails.env == 'production' ? 'www' : 'sandbox'}\"
+    data-currency=\"EUR\"
+    data-env=\"#{Rails.env == 'production' ? '' : 'sandbox'}\"
     data-callback=\"#{notify_url}\"
     data-return=\"#{return_url}\"
     ></script>".html_safe
