@@ -2,45 +2,33 @@ Airesis::Application.configure do
   # Settings specified here will take precedence over those in config/application.rb
 
   # In the development environment your application's code is reloaded on
-  # every request.  This slows down response time but is perfect for development
+  # every request. This slows down response time but is perfect for development
   # since you don't have to restart the web server when you make code changes.
   config.cache_classes = false
 
-  # Log error messages when you accidentally call methods on nil.
-  config.whiny_nils = true
+  # Do not eager load code on boot.
+  config.eager_load = false
 
-  # Show full error reports and disable caching
-  config.consider_all_requests_local       = false
+  # Show full error reports and disable caching.
+  config.consider_all_requests_local = false
   config.action_controller.perform_caching = false
 
-  # Print deprecation notices to the Rails logger
+  # Print deprecation notices to the Rails logger.
   config.active_support.deprecation = :log
 
-  # Only use best-standards-support built into browsers
-  config.action_dispatch.best_standards_support = :builtin
+  # Raise an error on page load if there are pending migrations
+  config.active_record.migration_error = :page_load
 
-
-  # Raise exception on mass assignment protection for Active Record models
-  config.active_record.mass_assignment_sanitizer = :strict
-
-# Log the query plan for queries taking more than this (works
-# with SQLite, MySQL, and PostgreSQL)
-  config.active_record.auto_explain_threshold_in_seconds = 1
+  config.assets.debug = true
 
   config.i18n.fallbacks = true
-
-  # Do not compress assets
-  config.assets.compress = false
-
-  # Expands the lines which load the assets
-  config.assets.debug = true
-  
+    
   #config.assets.logger = nil
   config.force_ssl = false
   
   config.action_mailer.delivery_method = :smtp
   config.action_mailer.perform_deliveries = true
-  config.action_mailer.raise_delivery_errors = true
+  config.action_mailer.raise_delivery_errors = false
   config.action_mailer.default_url_options = { :host => MAILER_DEFAULT_HOST }
   
   
@@ -90,28 +78,5 @@ ActionMailer::Base.smtp_settings = {
     :user_name          => EMAIL_USERNAME,
     :password           => EMAIL_PASSWORD
 }
-
-# Use this hook to configure devise mailer, warden hooks and so forth. The first
-# four configuration values can also be set straight in your models.
-Devise.setup do |config|
-  require "omniauth-facebook"
-  config.omniauth :facebook, FACEBOOK_APP_ID, FACEBOOK_APP_SECRET,
-                      {:scope => 'email', :client_options => {:ssl => {:verify => false, :ca_path => '/etc/ssl/certs'}}}
-                      
-  require "omniauth-google-oauth2"
-  config.omniauth :google_oauth2, GOOGLE_APP_ID, GOOGLE_APP_SECRET, { access_type: "offline", approval_prompt: "" }
-
-  require "omniauth-twitter"
-  config.omniauth :twitter,TWITTER_APP_ID, TWITTER_APP_SECRET
-
-  require "omniauth-meetup"
-  config.omniauth :meetup,MEETUP_APP_ID, MEETUP_APP_SECRET
-
-  require "omniauth-linkedin"
-  config.omniauth :linkedin, LINKEDIN_APP_ID, LINKEDIN_APP_SECRET
-
-  require "omniauth-parma"
-  config.omniauth :parma, PARMA_APP_ID, PARMA_APP_SECRET, {:scope => 'email basic'}
-end
 
 Airesis::Application.default_url_options = Airesis::Application.config.action_mailer.default_url_options
