@@ -1,1 +1,1 @@
-VERSION = '3.2'
+VERSION = '3.2.2'
